+++ Gemfile
@@ -130,13 +130,14 @@ group :development do
   # gem "spring"
 end
 
+gem "faker"
+
 group :test do
   # Use system testing [https://guides.rubyonrails.org/testing.html#system-testing]
   gem "capybara"
   gem "selenium-webdriver"
   gem "webdrivers"
   gem "shoulda-matchers", "~> 5.0"
-  gem "faker"
   gem "database_cleaner"
   gem 'database_cleaner-active_record'
 end
