+++ app/controllers/user_invitations_controller.rb
@@ -1,18 +1,41 @@
 class UserInvitationsController < ApplicationController
   before_action :authenticate_user!
 
+  def index
+    @invitations = User.where(invited_by_id: current_user.id)
+    respond_to do |format|
+      format.json
+    end
+  end
+
   def create
     if current_user.has_invitations_left?
-      user = User.invite!({email: params[:email]}, current_user)
+      user = User.invite!({username: Faker::Internet.username, email: params[:invitation][:email]}, current_user)
       if user.valid?
-        # user.update(role: "artist")
         current_user.decrement(:invitations_count)
         current_user.save
-        flash.now[:notice] = "User invited"
+        
+        @invitation = user
+        respond_to do |format|
+          format.json { render :create }
+        end
+      else
+        respond_to do |format|
+          format.json { 
+            render json: { 
+              error: user.errors.full_messages.join(", ")
+            }, status: :unprocessable_entity 
+          }
+        end
       end
-      @user = current_user
-      @section = "invitations"
-      render "user_settings/update"
+    else
+      respond_to do |format|
+        format.json { 
+          render json: { 
+            error: "No invitations left"
+          }, status: :unprocessable_entity 
+        }
+      end
     end
   end
 end
